+++ app/jobs/head.js
@@ -1,0 +1,10 @@
+ export default async function Head({ params }) {
+   
+    return (
+      <>
+        <title>Jobs | Rehacker - HackerNews NextJS Clone</title>
+        <link rel="canonical" href="/jobs" precedence="true" />
+      </>
+    )
+
+}