+++ components/posts/SinglePost.js
@@ -21,7 +21,6 @@ export default function SinglePost( props ) {
         setIsLoading(0);
     }, [isLoading, data, newsItem]);
 
-
     return (
         <>
             { isLoading == 1 ? (
@@ -37,14 +36,13 @@ export default function SinglePost( props ) {
                     { newsItem &&
                         <>
                             <div className="bg-white hover:bg-slate-100 transition ease-in-out delay-150 shadow-xl shadow-gray-100 w-full max-w-5xl px-5 py-4 border-b-2 border-grey" data-itemtype={ newsItem.type }>
-                                <div className="grid grid-cols-10">
-                                    <div className="col-span-1">
-                                        <div className="bg-hn-orange text-white rounded-full px-3 py-1 text-sm w-4/5 text-center pt-4 pb-4">
-                                            <span className="block text-xl w-full text-center">{ newsItem.score }</span>
-                                        </div>
+                                <div className="grid grid-cols-10 items-center">
+                                    <div className="hidden lg:block col-span-1">
+                                        <svg xmlns="http://www.w3.org/2000/svg" className="h-4 w-4 fill-hn-orange inline-block" viewBox="0 0 576 512"><path d="M316.9 18C311.6 7 300.4 0 288.1 0s-23.4 7-28.8 18L195 150.3 51.4 171.5c-12 1.8-22 10.2-25.7 21.7s-.7 24.2 7.9 32.7L137.8 329 113.2 474.7c-2 12 3 24.2 12.9 31.3s23 8 33.8 2.3l128.3-68.5 128.3 68.5c10.8 5.7 23.9 4.9 33.8-2.3s14.9-19.3 12.9-31.3L438.5 329 542.7 225.9c8.6-8.5 11.7-21.2 7.9-32.7s-13.7-19.9-25.7-21.7L381.2 150.3 316.9 18z"/></svg>
+                                        <span className="inline-block text-xl pl-2 relative top-[3px]">{ newsItem.score }</span>
                                     </div>
 
-                                    <div className="col-span-9">
+                                    <div className="col-span-10 lg:col-span-9 items-center">
                                         <h3 className="font-bold font-base mt-px">
                                             { newsItem.url ? (
                                                 <>
@@ -60,20 +58,22 @@ export default function SinglePost( props ) {
                                             <span className="font-normal text-xs pl-2"><DomainOutput url={ newsItem.url } /></span>
                                         </h3>
 
-                                        <div className="items-center">                                    
-                                            <span className="text-slate-600 text-xs items-center pr-2"> 
-                                                by <Link href="/user/[username]" as={`/user/${newsItem.by}`} className="hover:underline underline-offset-2">{ newsItem.by }</Link>
+                                        <div className="items-center">             
+                                            <div className="flex gap-2">
+                                                <div className="col-span-1 text-xs">
+                                                    Created by <Link href="/user/[username]" as={`/user/${newsItem.by}`} className="hover:underline underline-offset-2">{ newsItem.by }</Link>
+                                                </div>
 
-                                                <span className="pl-[5px] text-xs">
-                                                    | <DateOutput time={ newsItem.time } />
-                                                </span>
+                                                <div className="col-span-1 text-xs">
+                                                    <DateOutput time={ newsItem.time } />
+                                                </div>
 
-                                                <span className="pl-2 text-xs">
-                                                    | <Link href="/item/[id]" as={`/item/${newsItem.id}`} className="hover:underline underline-offset-2 text-xs">        
+                                                <div className="col-span-1 text-xs">
+                                                    <Link href="/item/[id]" as={`/item/${newsItem.id}`} className="hover:underline underline-offset-2 text-xs">        
                                                         { newsItem.descendants } comments
                                                     </Link>
-                                                </span>
-                                            </span>
+                                                </div>
+                                            </div>                                            
                                         </div>
                                     </div>
                                 </div>